fullstack/backend/pyproject.toml
[project]
name = "docusaurus-rag-ingestion"
version = "0.1.0"
description = "RAG ingestion pipeline for Docusaurus-based book content"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "cohere>=5.0.0",
    "qdrant-client>=1.8.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
    "openai>=1.0.0",
    "openai-agents>=0.6.4",
    "agents>=0.0.7",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
